[Bug, Refactor] SSFilter 모듈 생성 및 Sent and Received Module 리팩토링 , FilterLayout 버그 수정 (#622)
* prj: ssFilter생성

* feat: SSFilter Dependencies 추가,

* feat: SSFilterDateProperty 작성 SSFilterItemable 작성 Reducer 및 뷰 생성

* feat: dateWithFilter Reducer생성

* feat: sliderPropertyReducer 생성

* feat: sliderProperty와 연결 FilterReducer연결

* feat: sliderProperty 추가

* feat: filterView 완성

* feat: SSFilterView 연결

* feat: sent Module SSFilter적용

* Refactor: 필요 없는 파일 삭제

* feat: LedgerDetailFilterView and Reducer Refactor

* feat: Recived Filter 리팩토링

* feat: receivedFilter and RedciveFilterView 리팩토링

* feat: receivedFilter과 관련된 코드 리팩토링

* feat: receivedMainFilter 초기값 세팅 로직 생성

* feat: sliderProperty를 활용한 초기값 세팅 로직 작성

* feat: update filter initial State position

* feat: SentEnvelopeFilter initialPrev Value update

+++ Projects/Feature/Sent/Sources/SentMainView/SentMainView.swift
@@ -142,7 +142,7 @@ struct SentMainView: View {
             buttonText: person.name
           )
         ) {
-          store.sendViewAction(.tappedFilteredPersonButton(id: person.id))
+          store.sendViewAction(.tappedFilteredPersonButton(id: Int64(person.id)))
         }
       }
     }
